backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
mysqlclient>=2.2,<3.0
asyncua>=1.1,<2.0
